+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const QUALITY: u8 = 75;
+pub const QUALITY: i32 = 75;
 pub const PATTERNS: &str = "*.{jpg,jpeg,png}";
